Pong Gameplay changes
A few changes to the gameplay of Pong. I have increased the speed of the Paddles from 2 to 3 in both the up and down direction, making them move faster with the keys. The background colour of the board has been changed to Blue from Black. Players now need 5 points to win, as opposed to the 3 they needed before. The text that is displayed when there is a winner has been changed from "WIN!" to "WINNER WINNER CHICKEN DINNER!". The alignment for this will need to be tested. With the increased length, this could look off.